Merge interactive plot to other plotting method
To avoid any duplicate code,
plot_coarse-fine_vs_image_interactiveVersion.py
was removed and an option flag was added to
plot_coarse-fine_vs_image.py
This allows the user to have a pop-up window showing the desired plot.
In the default config parameter, an interactive key is set for activating
the interactive plotting method.