Correction of correction formula and add argparse
Modification of way of correcting coarse values to be independent from
slope convention: according to RAL the slope has to be considered
positive and not negative. For both sample and reset, the absolute
value of the slope is considered.

Add argument passer to avoid of modifying script for using it