WIP: Correct bug with Vin and gather informations
This is a work in progress for trying to avoid duplicate data like vin.
Instead, the directory path of the gathered data used during processing is store
in the collection section of the processed filed, depending of the adc_part.

For the characterization, this filename is used to load gathered data (data + vin)
but this is not working well properly (only for one adc part)